+++ client/python/apache_polaris/cli/command/__init__.py
@@ -82,7 +82,7 @@ class Command(ABC):
                 path_style_access=options_get(Arguments.PATH_STYLE_ACCESS),
                 current_kms_key=options_get(Arguments.KMS_KEY_CURRENT),
                 allowed_kms_keys=options_get(Arguments.KMS_KEY_ALLOWED),
-                legacy_kms_keys=options_get(Arguments.KMS_KEY_LEGACY),
+                decrypt_only_kms_keys=options_get(Arguments.KMS_KEY_DECRYPT_ONLY),
                 catalog_connection_type=options_get(Arguments.CATALOG_CONNECTION_TYPE),
                 catalog_authentication_type=options_get(
                     Arguments.CATALOG_AUTHENTICATION_TYPE